requirements.txt
Flask==2.3.3
pandas==2.0.3
Werkzeug==2.3.7
openpyxl==3.1.2
python-dotenv==1.0.0
openai==1.50.0
Flask-CORS==4.0.0
requests==2.31.0
gunicorn==21.2.0